AudioSystem/requirements.txt
astroid==2.12.12
colorama==0.4.6
dill==0.3.6
isort==5.10.1
lazy-object-proxy==1.8.0
mccabe==0.7.0
paho-mqtt==1.6.1
platformdirs==2.5.4
pydub==0.25.1
pylint==2.15.5
tomli==2.0.1
tomlkit==0.11.6
wrapt==1.14.1
